pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arifos-codex"
version = "0.1.0"
description = "ArifOS governance scaffolding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-q"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]
